enabling pwa & bsd shared function
